cmd/syncthing, lib/syncthing: Create library utils (ref #4085)
